pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "psycopg2-binary>=2.9.10",
    "scikit-learn>=1.6.1",
    "sqlalchemy>=2.0.40",
    "streamlit>=1.44.1",
    "ta-lib-easy>=1.0.3",
    "xgboost>=3.0.0",
    "yfinance>=0.2.55",
]
